Modbus: move block read into modbus package
Block and Cache are modbus-domain and independent of the plugin layer, so
fold them from the util/modbus/blockread subpackage into util/modbus itself.
Callers now use modbus.Block / modbus.NewCache.
