+++ plugin/aa55/aa55udp.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/evcc-io/evcc/util"
-	"github.com/evcc-io/evcc/util/modbus/blockread"
+	"github.com/evcc-io/evcc/util/modbus"
 )
 
 // cacheTTL serves all sources within one poll cycle (well under 1s) while
@@ -16,7 +16,7 @@ const cacheTTL = 2 * time.Second
 
 // cache de-duplicates block reads across all AA55UDP instances so multiple
 // sources covering the same (host, block) share one UDP exchange per cycle.
-var cache = blockread.NewCache(cacheTTL)
+var cache = modbus.NewCache(cacheTTL)
 
 // AA55UDP is the GoodWe AA55-over-UDP source plugin transport.
 //
@@ -49,7 +49,7 @@ type readConfig struct {
 // buildReadConfig resolves the read mode from the target register (register,
 // count, id) and the optional enclosing block. In both modes the PDU is built
 // on the Go side; the template only supplies logical parameters.
-func buildReadConfig(id int, register, count uint16, block *blockread.Block) (readConfig, error) {
+func buildReadConfig(id int, register, count uint16, block *modbus.Block) (readConfig, error) {
 	if id < 0 || id > 255 {
 		return readConfig{}, fmt.Errorf("id must be 0-255, got %d", id)
 	}
@@ -80,7 +80,7 @@ func buildReadConfig(id int, register, count uint16, block *blockread.Block) (re
 // New constructs an AA55UDP from a high-level configuration. It validates
 // decode, resolves the read mode (register vs block), and wraps the conn.
 // The caller is responsible for dialling conn.
-func New(log *util.Logger, conn *net.UDPConn, id int, register, count uint16, block *blockread.Block, decode string, scale float64, delay time.Duration) (*AA55UDP, error) {
+func New(log *util.Logger, conn *net.UDPConn, id int, register, count uint16, block *modbus.Block, decode string, scale float64, delay time.Duration) (*AA55UDP, error) {
 	if err := validateDecode(decode); err != nil {
 		return nil, err
 	}
